MSK21QAS-15 [board] moved context menu component folder
